+++ maa.go
@@ -10,8 +10,6 @@ import (
 var (
 	inited bool
 
-	ErrAlreadyInitialized     = errors.New("maa framework already initialized")
-	ErrNotInitialized         = errors.New("maa framework not initialized")
 	ErrSetLogDir              = errors.New("failed to set log directory")
 	ErrSetSaveDraw            = errors.New("failed to set save draw option")
 	ErrSetStdoutLevel         = errors.New("failed to set stdout level")
@@ -142,11 +140,12 @@ func WithJSONDecoder(decoder JSONDecoder) InitOption {
 
 // Init loads the dynamic library related to the MAA framework and registers its related functions.
 // It must be called before invoking any other MAA-related functions.
+// It must not be called concurrently with Release or other MAA-related functions.
 // Note: If this function is not called before other MAA functions, it will trigger a null pointer panic.
 func Init(opts ...InitOption) error {
 
 	if inited {
-		return ErrAlreadyInitialized
+		return nil
 	}
 
 	cfg := initConfig{}
@@ -155,14 +154,14 @@ func Init(opts ...InitOption) error {
 		opt(&cfg)
 	}
 
-	if err := native.Init(cfg.LibDir); err != nil {
+	if err := native.Initialize(cfg.LibDir); err != nil {
 		return err
 	}
 
 	success := false
 	defer func() {
 		if !success {
-			_ = native.Release()
+			_ = native.Shutdown()
 		}
 	}()
 
@@ -209,19 +208,16 @@ func Init(opts ...InitOption) error {
 }
 
 // IsInited checks if the MAA framework has been initialized.
+// It must not be called concurrently with Init or Release.
 func IsInited() bool {
 	return inited
 }
 
 // Release releases the dynamic library resources of the MAA framework and unregisters its related functions.
-// It must be called only after the framework has been initialized via Init.
+// It must not be called concurrently with Init or other MAA-related functions.
 func Release() error {
 
-	if !inited {
-		return ErrNotInitialized
-	}
-
-	if err := native.Release(); err != nil {
+	if err := native.Shutdown(); err != nil {
 		return err
 	}